Split `slate-lists` lib folder into smaller ones
