+++ packages/slate-lists/src/index.ts
@@ -1,4 +1,5 @@
 export * as ListsEditor from './lib';
+export * as Normalizations from './normalizations';
 export * from './types';
 export { withLists } from './withLists';
 export { withListsReact } from './withListsReact';
