+++ packages/slate-lists/src/lib/index.ts
@@ -1,5 +1,4 @@
 export { canDeleteBackward } from './canDeleteBackward';
-export { cloneContentsMonkeyPatch } from './cloneContentsMonkeyPatch';
 export { decreaseDepth } from './decreaseDepth';
 export { decreaseListItemDepth } from './decreaseListItemDepth';
 export { getListItemsInRange } from './getListItemsInRange';
@@ -16,14 +15,6 @@ export { listItemContainsText } from './listItemContainsText';
 export { mergeListWithPreviousSiblingList } from './mergeListWithPreviousSiblingList';
 export { moveListItemsToAnotherList } from './moveListItemsToAnotherList';
 export { moveListToListItem } from './moveListToListItem';
-export { normalizeList } from './normalizeList';
-export { normalizeListChildren } from './normalizeListChildren';
-export { normalizeListItemChildren } from './normalizeListItemChildren';
-export { normalizeListItemTextChildren } from './normalizeListItemTextChildren';
-export { normalizeOrphanListItem } from './normalizeOrphanListItem';
-export { normalizeOrphanListItemText } from './normalizeOrphanListItemText';
-export { normalizeOrphanNestedList } from './normalizeOrphanNestedList';
-export { normalizeSiblingLists } from './normalizeSiblingLists';
 export { setListType } from './setListType';
 export { splitListItem } from './splitListItem';
 export { unwrapList } from './unwrapList';
