+++ packages/slate-lists/src/normalizations/normalizeList.ts
