+++ packages/slate-lists/src/normalizations/normalizeListChildren.ts
