+++ packages/slate-lists/src/normalizations/normalizeListItemChildren.ts
