+++ packages/slate-lists/src/normalizations/normalizeListItemTextChildren.ts
