+++ packages/slate-lists/src/util/index.ts
@@ -1,0 +1,2 @@
+export { patchRangeCloneContents } from './patchRangeCloneContents';
+export { withRangeCloneContentsPatched } from './withRangeCloneContentsPatched';
