+++ packages/slate-lists/src/util/withRangeCloneContentsPatched.ts
@@ -1,0 +1,10 @@
+import { patchRangeCloneContents } from './patchRangeCloneContents';
+
+export function withRangeCloneContentsPatched(callback: () => void) {
+    const undo = patchRangeCloneContents();
+    try {
+        callback();
+    } finally {
+        undo();
+    }
+}
